// src/app/types/cartItemType.ts
export interface cartItem {
	id: number;
	product_code: string;
	product_title_en: string;
	product_title_bn: string;
	current_prices: number;
	regular_prices: number;
	quantity: number;
	discount_amount: number;
	product_image: string;
	product_weight: number;
	product_weight_type: string;
	pack_size: string;
}